dev: make adding new card
