+++ index.html
@@ -85,7 +85,7 @@
       </li>
     </template>
   </div>
-  <script type="module" src="js/script.js"></script>
+  <script type="module" src="js/index.js"></script>
   <script src="js/validation.js"></script>
 </body>
 </html>
